drivers:iio:stm:imu:st_lsm6dsvx,st_lsm6dsv16bx: fix wrong sign in odr calibration
The value of the INTERNAL_FREQ_FINE register represent the
difference in percentage of the effective ODR with respect
to the typical in a 8-bit format, two's complement.
Currently it is erroneously interpreted as an unsigned integer.

+++ drivers/iio/stm/imu/st_lsm6dsvx/st_lsm6dsvx_core.c
@@ -430,7 +430,7 @@ static int st_lsm6dsvx_get_odr_calibration(struct st_lsm6dsvx_hw *hw)
 		return err;
 	}
 
-	odr_calib = (data * ST_LSM6DSVX_TS_CAL_COEFF) / 1000;
+	odr_calib = ((s8)data * ST_LSM6DSVX_TS_CAL_COEFF) / 1000;
 	hw->ts_delta_ns = ST_LSM6DSVX_TS_DELTA_NS - odr_calib;
 
 	dev_info(hw->dev, "Freq Fine %lld (ts %lld)\n",
